+++ src/events/voice-state-update.ts
@@ -1,9 +1,8 @@
 import { VoiceState } from "discord.js";
-import { Valentine, VoiceOnDemandTools } from "../programs";
+import { VoiceOnDemandTools } from "../programs";
 
 class VoiceStateUpdate {
   constructor(oldState: VoiceState, newState: VoiceState) {
-    Valentine.valentineVoiceState(oldState, newState);
     VoiceOnDemandTools.voiceOnDemandPermissions(oldState, newState);
     VoiceOnDemandTools.voiceOnDemandReset(oldState, newState);
   }
