+++ src/games/configurator-validator.ts
