+++ src/games/index.ts
@@ -1,2 +1,2 @@
-export { default as GameHub } from "./GameHub";
-export { Spyfall } from "./spyfall/Spyfall";
+export { default as GameHub } from "./game-hub";
+export { Spyfall } from "./spyfall/spyfall";
