+++ src/programs/index.ts
@@ -1,46 +1,41 @@
-import * as BirthdayManagerTools from "./BirthdayManager";
-import * as DailyChallenge from "./DailyChallenge";
-import * as DMMenu from "./DMMenu";
-import * as Game from "./Game";
-import * as GroupManagerTools from "./GroupManager";
-import * as MapTools from "./Map";
-import * as NitroColors from "./NitroColors";
-import * as Separators from "./Separators";
-import * as TopicManager from "./TopicManager";
-import * as Unassigned from "./Unassigned";
-import * as Valentine from "./Valentine";
-import * as VoiceOnDemandTools from "./VoiceOnDemand";
-import * as WhereAreYouFromManager from "./WhereAreYouFromManager";
-import AdventureGame from "./AdventureGame";
-import BirthdayManager from "./BirthdayManager";
-import Deadchat from "./Deadchat";
-import EasterEvent from "./EasterEvent";
-import ExportManager from "./ExportManager";
-import GroupManager from "./GroupManager";
-import PollsManager from "./PollsManager";
-import ProfileManager from "./ProfileManager";
-import ReactRole from "./ReactRole";
-import Resource from "./ResourceManager";
-import Someone from "./Someone";
-import TemplateMode from "./TemplateMode";
-import Ticket from "./Ticket";
-import VoiceOnDemand from "./VoiceOnDemand";
+import * as BirthdayManagerTools from "./birthday-manager";
+import * as DailyChallenge from "./daily-challenge";
+import * as DMMenu from "./dm-menu";
+import * as Game from "./game";
+import * as GroupManagerTools from "./group-manager";
+import * as MapTools from "./map";
+import * as NitroColors from "./nitro-colors";
+import * as Separators from "./separators";
+import * as TopicManager from "./topics";
+import * as Unassigned from "./unassigned";
+import * as VoiceOnDemandTools from "./voice-on-demand";
+import * as WhereAreYouFrom from "./where-are-you-from";
+import BirthdayManager from "./birthday-manager";
+import Deadchat from "./deadchat";
+import ExportManager from "./export-manager";
+import GroupManager from "./group-manager";
+import Polls from "./polls";
+import Profile from "./profile";
+import ReactRole from "./react-role";
+import Resource from "./resources";
+import Someone from "./someone";
+import TemplateMode from "./template-mode";
+import Ticket from "./ticket";
+import VoiceOnDemand from "./voice-on-demand";
 
 export {
-  AdventureGame,
   BirthdayManager,
   BirthdayManagerTools,
   Deadchat,
   DMMenu,
-  EasterEvent,
   ExportManager,
   Game,
   GroupManager,
   GroupManagerTools,
   MapTools,
   NitroColors,
-  PollsManager,
-  ProfileManager,
+  Polls,
+  Profile,
   ReactRole,
   Resource,
   DailyChallenge,
@@ -50,8 +45,7 @@ export {
   Ticket,
   TopicManager,
   Unassigned,
-  Valentine,
   VoiceOnDemand,
   VoiceOnDemandTools,
-  WhereAreYouFromManager,
+  WhereAreYouFrom,
 };
